begin implementation - working for static images
We have updated sctk to version 0.17.0. Alongside that, we are now
currently using glutin for rendering directly with the gpu. This allows
us to store a buffer with just `rgb` values, as opposed to before, where
we stored `bgra` values.

Furthermore, because the buffer is no longer memmmapped, and we have
total control over it, we can wrap it in an Arc and make it readable
across threads. This will let us avoid a nasty `clone` we were forced to
do before, and completely eliminate the `ReadyPack` struct.

Finally, it should also make all functions that looped over all pixels
faster, since they will have less bytes to worry about. This includes
the compression/decompression functions and the transition functions.